requirements.txt
allure-pytest
Appium-Python-Client
paramiko
pika
pipdeptree
pipreqs
pytest-csv
pytest-rerunfailures
pytest-sugar
pytest-xdist
reportlab
sshtunnel
tabulate
toml
tqdm
